Extract block forger
