+++ packages/validator/source/service-provider.ts
@@ -7,6 +7,7 @@ import { Providers } from "@mainsail/kernel";
 import { assert } from "@mainsail/utils";
 import Joi from "joi";
 
+import { BlockForger } from "./block-forger.js";
 import { BIP38, BIP39 } from "./keys/index.js";
 import { TransactionForger } from "./transaction-forger.js";
 import { ValidatorRepository } from "./validator-repository.js";
@@ -17,6 +18,7 @@ export class ServiceProvider extends Providers.ServiceProvider {
 	public async register(): Promise<void> {
 		this.app.bind(Identifiers.Validator.Repository).to(ValidatorRepository).inSingletonScope();
 		this.app.bind(Identifiers.Validator.TransactionForger).to(TransactionForger).inSingletonScope();
+		this.app.bind(Identifiers.Validator.BlockForger).to(BlockForger).inSingletonScope();
 	}
 
 	public async boot(): Promise<void> {
